requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
matplotlib>=3.9
scipy>=1.11
joblib>=1.4
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.8
PyYAML>=6.0
pytest>=8.0
# optional
torch>=2.5
xgboost>=2.0
lightgbm>=4.3
shap>=0.46
